refactor: [#1195] move tests to KeysHandler
These tests do not require the authentication service.
